Revert "Svelte Template Format"
