+++ .prettierrc
@@ -1,4 +1,3 @@
 {
-  "endOfLine": "auto",
-  "pluginSearchDirs": ["."]
+  "endOfLine": "auto"
 }
