+++ package.json
@@ -9,26 +9,29 @@
     "build": "eleventy",
     "clean": "rimraf build",
     "dev": "eleventy --serve",
-    "format:check": "prettier --check --ignore-path .gitignore \"**/*.{js,svelte,json,yaml,scss,css}\"",
-    "format": "prettier --write --ignore-path .gitignore \"**/*.{js,svelte,json,yaml,scss,css}\""
+    "format:check": "prettier --check --ignore-path .gitignore \"**/*.{js,jsx,json,yaml,scss,css}\"",
+    "format": "prettier --write --ignore-path .gitignore \"**/*.{js,jsx,json,yaml,scss,css}\""
   },
   "devDependencies": {
-    "prettier": "^2.6.2",
-    "prettier-plugin-svelte": "^2.6.0",
-    "rimraf": "^3.0.2"
-  },
-  "dependencies": {
     "@11ty/eleventy": "^1.0.0",
     "@11ty/eleventy-plugin-syntaxhighlight": "^4.0.0",
+    "@babel/core": "^7.17.8",
+    "@babel/preset-env": "^7.16.11",
+    "@babel/preset-react": "^7.16.7",
+    "@babel/register": "^7.17.7",
     "@iktakahiro/markdown-it-katex": "^4.0.1",
     "eleventy-plugin-nesting-toc": "^1.3.0",
     "fast-xml-parser": "^4.0.7",
     "feed": "^4.2.2",
-    "got": "^12.0.3",
+    "got": "^12.0.2",
     "js-yaml": "^4.1.0",
     "markdown-it": "^12.3.2",
-    "markdown-it-anchor": "^8.5.0",
-    "sass": "^1.49.11",
-    "svelte": "^3.46.6"
-  }
+    "markdown-it-anchor": "^8.4.1",
+    "preact": "^10.6.6",
+    "preact-render-to-string": "^5.1.20",
+    "prettier": "^2.6.0",
+    "rimraf": "^3.0.2",
+    "sass": "^1.49.9"
+  },
+  "dependencies": {}
 }
