+++ src/_includes/BlogHeader.jsx
@@ -1,0 +1,9 @@
+export const BlogHeader = () => (
+  <header class="site-header">
+    <div class="title">
+      <a href="/blog/" class="_reset-a">
+        Honai's Blog
+      </a>
+    </div>
+  </header>
+);
