--- src/_includes/BlogHeader.svelte
@@ -1,5 +1,0 @@
-<header class="site-header">
-  <div class="title">
-    <a href="/blog/" class="_reset-a"> Honai's Blog </a>
-  </div>
-</header>
