+++ src/_includes/Footer.jsx
@@ -1,0 +1,17 @@
+export const Footer = () => (
+  <footer class="site-footer">
+    &copy; 2021 Honai Ueoka
+    <br />
+    <a href="/" class="_uncolor">
+      Portfolio
+    </a>
+    &ensp;
+    <a href="/works/" class="_uncolor">
+      Works
+    </a>
+    &ensp;
+    <a href="/blog/" class="_uncolor">
+      Blog
+    </a>
+  </footer>
+);
