+++ src/_includes/Script.jsx
@@ -1,0 +1,3 @@
+export const Script = ({ children, ...attrs }) => (
+  <script {...attrs} dangerouslySetInnerHTML={{ __html: children }} />
+);
