--- src/index.svelte
@@ -1,193 +1,0 @@
-<script>
-  import { getJsFuncs } from "./_includes/SvelteContext.svelte";
-  import PortfolioLayout from "./_includes/PortfolioLayout.svelte";
-  export let profile;
-  export let feeds;
-  export let page;
-  const { isodate, svginline } = getJsFuncs();
-</script>
-
-<PortfolioLayout pageUrl={page.url}>
-  <div class="nav-title">
-    <h1 class="title">Hi👋 I'm Honai.</h1>
-    <nav class="nav">
-      <a href="/works">Works</a>
-      <a href="/blog/">Blog</a>
-    </nav>
-  </div>
-
-  <section class="simple-card">
-    <h2 class="heading" id="about">
-      <a href="#about" class="anchor"> About </a>
-    </h2>
-    <div class="content">
-      <div class="profile-section">
-        {@html svginline("asisai")}
-        <div class="name">
-          <div class="ja">ほない</div>
-          <div class="en">Honai Ueoka</div>
-        </div>
-      </div>
-    </div>
-    <div class="content">
-      <p>
-        京都大学大学院 情報学研究科 通信情報システム専攻 /
-        修士課程（2023年3月修了予定） /{" "}
-        <a href="https://camph.net/" target="_blank" rel="noopener">
-          CAMPHOR-
-        </a>
-        運営メンバー
-      </p>
-    </div>
-    <div class="content">
-      <ul class="social-links _reset-ul">
-        {#each profile.links as L}
-          <li class="item">
-            <a
-              href={L.url}
-              target="_blank"
-              rel="noopener"
-              class="link _reset-a"
-            >
-              {@html svginline(L.icon)}
-              <span class="text">{L.text}</span>
-            </a>
-          </li>
-        {/each}
-      </ul>
-    </div>
-  </section>
-
-  <section class="simple-card">
-    <h2 class="heading" id="education">
-      <a href="#education" class="anchor"> Education </a>
-    </h2>
-    {#each profile.education as E}
-      <div class="content">
-        <div class="education-item">
-          <div class="period">{E.period}</div>
-          <div class="desc">{E.title}</div>
-        </div>
-      </div>
-    {/each}
-  </section>
-
-  <section class="simple-card">
-    <h2 class="heading" id="intern">
-      <a href="#intern" class="anchor"> Internships / Employments </a>
-    </h2>
-    {#each profile.jobs as E}
-      <div class="content">
-        <div class="education-item">
-          <div class="period">{E.period}</div>
-          <div class="desc">{E.title}</div>
-        </div>
-      </div>
-    {/each}
-  </section>
-
-  <section class="simple-card">
-    <h2 class="heading" id="publications">
-      <a href="#publications" class="anchor"> Publications </a>
-    </h2>
-    {#each profile.publications as P}
-      <div class="content">
-        {P.authors}
-        <br />
-        <strong>{P.title}</strong>
-        <br />
-        {P.journal}
-        <br />
-        {#each P.links as L}
-          <a href={L.url} target="_blank" rel="noopener">
-            {L.text}
-          </a>
-          &ensp;
-        {/each}
-      </div>
-    {/each}
-  </section>
-
-  <section class="simple-card">
-    <h2 class="heading" id="posts">
-      <a href="#posts" class="anchor"> Posts </a>
-    </h2>
-    {#each feeds as post}
-      <div class="content post-item">
-        <a href={post.link} target="_blank" rel="noopener">
-          {post.title}
-        </a>
-        <div class="source">
-          <time dateTime={isodate(post.pubDate)}>
-            {isodate(post.pubDate)}
-          </time>
-          &middot;
-          <a
-            href={post.sourceUrl}
-            target="_blank"
-            rel="noopener"
-            class="_uncolor"
-          >
-            {post.sourceTitle}
-          </a>
-        </div>
-      </div>
-    {/each}
-  </section>
-
-  <section class="simple-card">
-    <h2 class="heading" id="presentations">
-      <a href="#presentations" class="anchor"> Presentations </a>
-    </h2>
-    <div class="content">
-      <ul class="_reset-ul slide-list">
-        {#each profile.slides as s}
-          <li class="slide-item">
-            <a href={s.url} class="link _reset-a">
-              <img
-                src={`/images/slide_thumb/${s.thumb}`}
-                alt={`${s.title}のスライドのサムネイル`}
-                width="640"
-                height="320"
-                class="thumb"
-                loading="lazy"
-              />
-              <div class="texts">
-                <div class="title">{s.title}</div>
-                <div class="subtitle">{s.subtitle}</div>
-                <div class="date">{s.date}</div>
-              </div>
-            </a>
-          </li>
-        {/each}
-      </ul>
-    </div>
-  </section>
-
-  <section class="simple-card">
-    <h2 class="heading" id="works">
-      <a href="#works" class="anchor"> Works </a>
-    </h2>
-    <div class="content">
-      <div class="works-gallery">
-        {#each profile.works.slice(0, 2) as w}
-          <a href={w.url} title={w.name} class="thumblink">
-            <img src={w.thumb} alt={w.name} class="thumb" loading="lazy" />
-          </a>
-        {/each}
-      </div>
-      <div class="works-page-link">
-        <a href="/works">See more works</a>
-      </div>
-    </div>
-  </section>
-
-  <section class="simple-card">
-    <h2 class="heading" id="blog">
-      <a href="#blog" class="anchor"> Blog </a>
-    </h2>
-    <div class="content">
-      <a href="/blog">Honai's Blog</a>
-    </div>
-  </section>
-</PortfolioLayout>
